feat: ensure delete task when project is deleted

+++ src/main/java/dev/ravel/proflow/domain/services/ProjectServiceImpl.java
@@ -1,6 +1,8 @@
 package dev.ravel.proflow.domain.services;
 
 import dev.ravel.proflow.infrastructure.model.Project;
+import dev.ravel.proflow.infrastructure.model.Subtask;
+import dev.ravel.proflow.infrastructure.model.Task;
 import dev.ravel.proflow.infrastructure.repository.ProjectRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,10 +12,12 @@ import java.util.List;
 @Service
 public class ProjectServiceImpl implements ProjectService {
     private final ProjectRepository projectRepository;
+    private final TaskService taskService;
 
     @Autowired
-    public ProjectServiceImpl(ProjectRepository projectRepository) {
+    public ProjectServiceImpl(ProjectRepository projectRepository, TaskService taskService) {
         this.projectRepository = projectRepository;
+        this.taskService = taskService;
     }
 
     @Override
@@ -38,7 +42,16 @@ public class ProjectServiceImpl implements ProjectService {
 
     @Override
     public void deleteProject(long id) {
-        projectRepository.deleteById(id);
+        Project project = projectRepository.findById(id).orElse(null);
+        if (project != null) {
+            List<Task> tasks = taskService.getTasksByProjectId((int) id);
+            if (tasks != null) {
+                tasks.forEach(task -> {
+                    taskService.deleteTask((int)task.getId());
+                });
+            }
+            projectRepository.deleteById(id);
+        }
     }
 
     @Override
